requirements.txt
googletrans==3.1.0a0
beautifulsoup4
requests
python-telegram-bot